(mod): continue changes in implementing assets properly.

+++ src/parser/parse_rgb.c
@@ -6,7 +6,7 @@
 /*   By: mattcarniel <mattcarniel@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/13 11:56:11 by mattcarniel       #+#    #+#             */
-/*   Updated: 2026/03/16 11:44:59 by mattcarniel      ###   ########.fr       */
+/*   Updated: 2026/03/17 11:57:37 by mattcarniel      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static int	parse_u8(t_str *s, unsigned char *out)
 
 	while (i < s->len && (s->ptr[i] >= '0' && s->ptr[i] <= '9'))
 	{
-		val = val * 10 + (s->ptr[i] - '0');
+		val = val * 10 + (uint32_t)(s->ptr[i] - '0');
 		if (val > 255)
 			return (1);
 		i++;
@@ -77,5 +77,5 @@ uint32_t	parse_rgb(t_str value)
 		return (0);
 	if (r == 0 && g == 0 && b == 0)
 		return (1); //return smallest color change value, reserve 0 black as error val
-	return (r << 16 | g << 8 | b);
+	return ((uint32_t)(r << 16 | g << 8 | b));
 }
